# src/main/resources/static/swagger.yaml
swagger: '2.0'
info:
  description: Описание тестового проекта
  version: '1.1'
  title: Тестовый проект
  termsOfService: Free to use
  contact:
    name: ADN
    url: 'https://vk.com/dimaaverchenko'
    email: 1434-3541@mail.ru
  license:
    name: Скачать проект
    url: 'https://github.com/PythonADN/Expense'
host: 'localhost:8080'
basePath: /
tags:
  - name: category-controller
    description: Category Controller
  - name: expense-controller
    description: Expense Controller
paths:
  /categories:
    get:
      tags:
        - category-controller
      summary: categories
      operationId: categoriesUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Category'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  /category:
    post:
      tags:
        - category-controller
      summary: createCategory
      operationId: createCategoryUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: category
          description: category
          required: true
          schema:
            $ref: '#/definitions/Category'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Category'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    put:
      tags:
        - category-controller
      summary: updateCategory
      operationId: updateCategoryUsingPUT
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: category
          description: category
          required: true
          schema:
            $ref: '#/definitions/Category'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Category'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/category/{id}':
    get:
      tags:
        - category-controller
      summary: categoryById
      operationId: categoryByIdUsingGET
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Category'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    delete:
      tags:
        - category-controller
      summary: deleteCategory
      operationId: deleteCategoryUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Category'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
  /expenses:
    get:
      tags:
        - expense-controller
      summary: getExpense
      operationId: getExpenseUsingGET
      produces:
        - '*/*'
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/Expense'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
    post:
      tags:
        - expense-controller
      summary: createExpenses
      operationId: createExpensesUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: expense
          description: expense
          required: true
          schema:
            $ref: '#/definitions/Expense'
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Expense'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
      deprecated: false
  '/expenses/{id}':
    delete:
      tags:
        - expense-controller
      summary: deleteExpense
      operationId: deleteExpenseUsingDELETE
      produces:
        - '*/*'
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Expense'
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
      deprecated: false
definitions:
  Category:
    type: object
    properties:
      id:
        type: integer
        format: int64
      name:
        type: string
    title: Category
  Expense:
    type: object
    properties:
      category:
        $ref: '#/definitions/Category'
      description:
        type: string
      expenseDate:
        type: string
        format: date-time
      id:
        type: integer
        format: int64
      location:
        type: string
    title: Expense